backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
scikit-learn==1.6.1
nltk==3.9.1
spacy>=3.8.0
pandas==2.2.3
numpy==2.2.3
joblib==1.4.2
Flask-Limiter==3.8.0
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
python-dotenv>=1.0.0
